Restrict WebAuthn bridge origins
